+++ dex/src/main/java/me/aki/tactical/dex/insn/litmath/RSubLitInstruction.java
@@ -12,7 +12,7 @@ import me.aki.tactical.dex.Register;
  * <code>result = op2 - op1;</code>
  */
 public class RSubLitInstruction extends AbstractBinaryLitMathInstruction {
-    public RSubLitInstruction(Register op1, int op2, Register result) {
+    public RSubLitInstruction(Register op1, short op2, Register result) {
         super(op1, op2, result);
     }
 }
